+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/edit/details/main.js
@@ -482,7 +482,7 @@ define([
             this.dfdListenForChange.then(function() {
                 this.sandbox.dom.on(constants.formSelector, 'change keyup', function() {
                     this.sandbox.emit('sulu.tab.dirty');
-                }.bind(this), 'select, input, textarea');
+                }.bind(this), 'select, input, textarea, .trigger-save-button');
 
                 // if a field-type gets changed or a field gets deleted
                 this.sandbox.on('sulu.contact-form.changed', function() {