+++ src/Sulu/Component/Websocket/MessageDispatcher/MessageDispatcherInterface.php
@@ -20,7 +20,7 @@ use Sulu\Component\Websocket\ConnectionContext\ConnectionContextInterface;
 interface MessageDispatcherInterface
 {
     /**
-     * @param string                  $name    Message name
+     * @param string $name Message name
      * @param MessageHandlerInterface $handler Handler which process messages with given name
      */
     public function add($name, MessageHandlerInterface $handler);
@@ -28,11 +28,25 @@ interface MessageDispatcherInterface
     /**
      * Dispatch event to handler with given name.
      *
-     * @param ConnectionInterface        $conn
-     * @param string                     $name    Message name
-     * @param array                      $message
-     * @param array                      $options
+     * @param ConnectionInterface $conn
+     * @param string $name Message name
+     * @param array $message
+     * @param array $options
      * @param ConnectionContextInterface $context
      */
-    public function dispatch(ConnectionInterface $conn, $name, array $message, array $options, ConnectionContextInterface $context);
+    public function dispatch(
+        ConnectionInterface $conn,
+        $name,
+        array $message,
+        array $options,
+        ConnectionContextInterface $context
+    );
+
+    /**
+     * All message handler get the message that the connection is losed.
+     *
+     * @param ConnectionInterface $conn
+     * @param ConnectionContextInterface $context
+     */
+    public function onClose(ConnectionInterface $conn, ConnectionContextInterface $context);
 }