+++ CHANGELOG.md
@@ -2,6 +2,14 @@ CHANGELOG for Sulu
 ==================
 
 * dev-develop
+    * BUGFIX      #1675 [ContactBundle]   Fixed null value for smart content
+    * FEATURE     #1653 [MediaBundle]     Added generation of thumbnails for videos
+    * BUGFIX      #1688 [ContentBundle]   Fixed doctrine cache size for preview with delete on navigate
+    * FEATURE     #1598 [MediaBundle]     Added infinite-scroll pagination for masonry-view
+    * FEATURE     #1598 [ContactBundle]   Added infinite-scroll pagination for card-view
+    * BUGFIX      #1670 [ContentBundle]   Fixed missing url-scheme in content type
+    * BUGFIX      #1673 [ContentBundle]   Fixed settings tag for shadows on external links
+    * BUGFIX      #1678 [ContactBundle]   Fixed contact-selection serialization
     * BUGFIX      #1667 [ContactBundle]   Removed the restriction of start dates from the datepicker
     * BUGFIX      #1671 [ContentBundle]   Fixed block sorting for blocks with only one type
     * BUGFIX      #1668 [ContentBundle]   Fixed smart content for usage with block property